requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
openpyxl==3.1.2
gunicorn==21.2.0
